All cards at home, cart, and favorites view are now going do productdetailsView with data
