+++ API-Challange/API-Challange/Components/ProductCardLarge.swift
@@ -78,5 +78,5 @@ struct ProductCardLarge: View {
 }
 
 //#Preview {
-//    ProductCardLarge(product: ProductModel(id: 2, title: "Sei la", description: "Loooooooonnnnnggg description", category: "quauqler coisa" , price: 60.00, discountPercentage: 30.00, thumbnail: ""))
+//    ProductCardLarge(product: ProductModel(id: 2, title: "Sei la", description: "Loooooooonnnnnggg description", category: "quauqler coisa" , price: 60.00, discountPercentage: 30.00, thumbnail: ""), toggleFavorite: )
 //}
