+++ API-Challange/API-Challange/Screens/FavoritesView.swift
@@ -15,7 +15,7 @@ struct FavoritesView: View {
     @State private var searchText = ""
     @State private var productNavigation: ProductModel = ProductModel(id: 0, title: "", description: "", category: "", price: 0, discountPercentage: 0, thumbnail: "")
 
-    var viewModel: FavoritesViewModel
+    @State var viewModel: FavoritesViewModel
     
     var body: some View {
         NavigationStack {
@@ -23,9 +23,10 @@ struct FavoritesView: View {
                 .navigationTitle("Favorites")
                 .searchable(text: $searchText)
                 .sheet(isPresented: $showCartSheet, content: {
-                    DetailView(product: productNavigation, toggleFavorite: {viewModel.serviceFavorites.toggleFavorite(productNavigation.id)})
+                    ProductDetailsView(product: productNavigation, toggleFavorite: {viewModel.toggleFavorite(productNavigation.id)})
                     .onDisappear {
                         Task {
+                            
                             await viewModel.loadingFavorites()
                             await viewModel.getFavoriteProducts()
                         }
