+++ API-Challange/API-Challange/ViewModels/Home/HomeViewModel.swift
@@ -20,15 +20,12 @@ final class HomeViewModel: HomeProtocol {
     
     var state: HomeState = .idle 
     private let serviceAPI: ProductAPIServiceProtocol
-    private var serviceFavorites: ProductFavoriteProtocol
+    var serviceFavorites: ProductFavoriteProtocol
     
     init(serviceAPI: ProductAPIServiceProtocol, serviceFavorites: ProductFavoriteProtocol) {
         self.serviceAPI = serviceAPI
         self.serviceFavorites = serviceFavorites
         
-//        self.serviceFavorites.onChange = {
-//            Task { await self.loadProducts() }
-//        }
     }
     
     func toggleFavorites(_ id: Int) {
