+++ API-Challange/API-Challange/ViewModels/ProductDetail/ProductDetailProtocol.swift
@@ -4,4 +4,7 @@
 //
 //  Created by Gustavo Ferreira bassani on 15/08/25.
 //
-
+//
+//protocol ProductDetailProtocol {
+//    
+//}
