+++ API-Challange/API-Challange/ViewModels/ProductDetail/ProductDetailViewModel.swift
@@ -5,14 +5,15 @@
 //  Created by Gustavo Ferreira bassani on 15/08/25.
 //
 
-import SwiftUI
-
-struct ProductDetailViewModel: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-#Preview {
-    ProductDetailViewModel()
-}
+//import SwiftUI
+//
+//struct ProductDetailViewModel: ProductDetailProtocol {
+//    
+//    private var serviceFavorites: ProductFavoriteProtocol
+//    
+//    init(serviceFavorites: ProductFavoriteProtocol) {
+//        self.serviceFavorites = serviceFavorites
+//    }
+//
+//    
+//}
